object cards in CreateClass Plan
